+++ docs/installation.rst
@@ -9,7 +9,7 @@ Installation
 Dependencies
 ------------
 
-Installing :code:`taurenmd` does **not** install **all** its dependencies. **Why?** Because we rely on large and complex dependencies required to manage the Molecular Dynamics (MD) data, such as :mda:`/` and :mdtraj:`/`, and we acknowledge that:
+Installing :code:`taurenmd` does **not** install **all** its dependencies. **Why?** Because we rely on large and complex dependencies required to manage the Molecular Dynamics (MD) data, such as `MDAnalysis <https://www.mdanalysis.org>`_ and `MDTraj <https://mdtraj.org/>`_, and we acknowledge that:
 
 1. a third party software (like *taurenmd*) should not attempt to guarantee proper installation of those complex (non-pure Python) libraries on its own (even via PyPi or Anaconda).
 2. many MD users actually do not work on the stable-release version of those *libs*, instead they work on:
@@ -21,7 +21,7 @@ Installing :code:`taurenmd` does **not** install **all** its dependencies. **Why
 3. for those who want to work with the stable-release version of these dependencies, their respective websites provide very straightforward ways to install them, you should refer to those.
 4. lastly and minor, not all dependencies are required for every *taurenmd command*.
 
-For these reasons, we have decided **not** to install these large dependencies together with *taurenmd*; other minor dependencies are automatically installed, though. Bellow a list of the dependencies you should install yourself prior to installating *taurenmd*.
+For these reasons, we have decided **not** to install these large dependencies together with *taurenmd*; other minor dependencies are automatically installed, though. Bellow a list of the dependencies you should install yourself prior to installing *taurenmd*.
 
 .. note::
     
@@ -40,10 +40,10 @@ Other dependencies automatically installed are listed bellow:
 #. `python-bioplottemplates <https://github.com/joaomcteixeira/python-bioplottemplates>`_
 #. `pyquaterion <http://kieranwynn.github.io/pyquaternion/>`_
 
-Compatible Platforms
---------------------
+Supported Platforms
+-------------------
 
-**taurenmd** is designed to run natively under any `platform compatible with Python <https://pythondev.readthedocs.io/platforms.html>`_ (paths are not hard coded :code:`;-)`). However, the libraries **taurenmd** depends on may or may not be compatible with all OS platforms, you should look for a platform compatible with all the required Molecular Dynamics analysis libraries. We can guarantee **taurenmd** works fully with all its dependencies under Ubuntu 18.04 LTS.
+**taurenmd** is designed to run natively under any `platform compatible with Python <https://pythondev.readthedocs.io/platforms.html>`_ (paths are not hard coded :code:`;-)`). However, the libraries **taurenmd** depends on may or may not be compatible with all OS platforms, you should look for a platform compatible with all the required Molecular Dynamics analysis libraries. We can guarantee **taurenmd** works fully with all its dependencies under Ubuntu 18.04 LTS. You may wish to read further on how continuous integration is :ref:`managed for this project <continuous integration>`.
 
 Installation steps
 ------------------
@@ -74,19 +74,29 @@ From scratch
 
     pip install taurenmd
 
+5. To update **taurenmd** simply::
+
+    pip install -upgrade taurenmd
+
 From GitHub
 ```````````
 
-If you are a proficient Pythonista you might want to install **taurenmd** from its *latest* source on GitHub. If that is the case you might not need to read this section because you know well what to do; nonetheless, let's go through it:
+If you are a proficient Pythonista you might want to install **taurenmd** from a development branch on GitHub. If that is the case you might not need to read this section because you know well what to do; nonetheless, let's go through it:
+
+.. note::
+
+    ``taurenmd`` follows Semantic Version 2.0, meaning that every single new addition to the master branch gets released on PyPI with a new version number.
+    Therefore, installing from the ``master`` GitHub branch actually adds no benefit to installing from PyPI.
 
 #. install the MD analysis libraries as described in the above sections
 #. clone our repository: :code:`git clone https://github.com/joaomcteixeira/taurenmd`
 #. place yourself in the new :code:`taurenmd` folder, in Linux-like systems: :code:`cd taurenmd`.
+#. ``git checkout the-branch-you-want-to-use``
 #. install **taurenmd** with the following command: :code:`python setup.py develop`
-#. in the future, to keep your installation up to the latests:
+#. in the future, to keep your installation up to the latest:
+
   #. pull repository updates from the upstream repository: :code:`git pull` (from within :code:`taurenmd` git folder)
   #. just in case something special was added, repeat :code:`python setup.py develop`
-
 
 .. _PyPi: https://pypi.org/
 .. _Anaconda: https://www.anaconda.com/distribution/
