+++ docs/reference/cli_angle.rst
@@ -1,0 +1,7 @@
+Client Angle
+============
+
+.. argparse::
+   :module: taurenmd.cli_angle
+   :func: _ap
+   :prog: tmdangle
