+++ docs/reference/cli_distances.rst
@@ -1,0 +1,7 @@
+Client Distances
+================
+
+.. argparse::
+   :module: taurenmd.cli_distances
+   :func: _ap
+   :prog: tmddist
