+++ docs/reference/cli_fext.rst
@@ -1,0 +1,7 @@
+Client Frame Extract
+====================
+
+.. argparse::
+   :module: taurenmd.cli_fext
+   :func: _ap
+   :prog: tmdfext
