+++ docs/reference/cli_imagemol.rst
@@ -1,0 +1,7 @@
+Client Image Molecule
+=====================
+
+.. argparse::
+   :module: taurenmd.cli_imagemol
+   :func: _ap
+   :prog: tmdimagemol
