+++ docs/reference/cli_nosol.rst
@@ -1,0 +1,7 @@
+Client No Solvent
+=================
+
+.. argparse::
+   :module: taurenmd.cli_nosol
+   :func: _ap
+   :prog: tmdnosol
