+++ docs/reference/cli_report.rst
@@ -1,0 +1,7 @@
+Client Report
+=============
+
+.. argparse::
+   :module: taurenmd.cli_report
+   :func: _ap
+   :prog: tmdreport
