+++ docs/reference/cli_rmsd.rst
@@ -1,0 +1,7 @@
+Client RMSD
+===========
+
+.. argparse::
+   :module: taurenmd.cli_rmsd
+   :func: _ap
+   :prog: tmdrmsd
