+++ docs/reference/cli_rmsf.rst
@@ -1,0 +1,7 @@
+Client RMSF
+===========
+
+.. argparse::
+   :module: taurenmd.cli_rmsf
+   :func: _ap
+   :prog: tmdrmsf
