+++ docs/reference/cli_rotations.rst
@@ -1,0 +1,7 @@
+Client Rotations
+================
+
+.. argparse::
+   :module: taurenmd.cli_rotations
+   :func: _ap
+   :prog: tmdrotations
