+++ docs/reference/cli_trajedit.rst
@@ -1,0 +1,7 @@
+Client Traj Edit
+================
+
+.. argparse::
+   :module: taurenmd.cli_trajedit
+   :func: _ap
+   :prog: tmdtrajedit
