+++ docs/reference/clients.rst
@@ -1,8 +1,28 @@
 Command-line interfaces
 =======================
 
-.. argparse::
-   :module: taurenmd.cli
-   :prog: taurenmd
-   :func: _ap
+This page documents all client interfaces available in ``taurenmd``, referred also as *taurenmd subroutines*. You may wish to read before the :ref:`Usage` page for general examples on how to use ``taurenmd`` for different purposes.
 
+
+You can access a list of all *taurenmd subroutines* by running on your terminal::
+
+    taurenmd
+
+taurenmd subroutines
+--------------------
+
+This TOC lists all *taurenmd subroutines*, click on each one to read on specific usage examples, and technical documentation.
+
+.. toctree::
+   :maxdepth: 1
+
+   cli_angle
+   cli_distances
+   cli_fext
+   cli_imagemol
+   cli_nosol
+   cli_report
+   cli_rmsd
+   cli_rmsf
+   cli_rotations
+   cli_trajedit
