+++ docs/reference/core.rst
@@ -1,0 +1,5 @@
+core
+====
+
+.. automodule:: taurenmd.core
+   :members:
