+++ docs/reference/index.rst
@@ -1,11 +1,12 @@
 Library Documentation
 =====================
 
-Documentation on the *command-line* interfaces is explained in `this page <Command-line interfaces>`_. Here the **taurenmd** internals are documented.
+Thorough documentation on the *command-line* interfaces exists in :ref:`this page <command-line interfaces>`. Here the **taurenmd** internals are documented, you may use them independently for a more advanced developer use.
+
+.. automodule:: taurenmd.libs
 
 .. toctree::
-    :maxdepth: 3
     :glob:
 
-    taurenmd
-    libs
+    core
+    lib*
