--- docs/reference/taurenmd.rst
@@ -1,5 +1,0 @@
-taurenmd module
-===============
-
-.. automodule:: taurenmd
-    :members:
