+++ docs/versioning.rst
@@ -1,0 +1,8 @@
+Versioning
+==========
+
+This project follows strictly `Semantic Versioning 2.0 <https://semver.org/#semantic-versioning-200>`_ for version control. 
+
+Upon release of version 1.0.0, all additions to the ``master`` branch are done by PR followed by its respective version increment and release on `PyPI <https://pypi.org/project/taurenmd/>`_.
+
+.. include:: ../CHANGELOG.rst
