+++ src/taurenmd/cli_fext.py
@@ -81,6 +81,10 @@ ap.add_argument(
     )
 
 
+def _ap():
+    return ap
+
+
 def main(
         topology,
         trajectories,
