+++ src/taurenmd/cli_rotations.py
@@ -64,6 +64,10 @@ libcli.add_plane_selection_arg(ap)
 libcli.add_slice_arg(ap)
 
 
+def _ap():
+    return ap
+
+
 def main(
         topology,
         trajectories,
