+++ tox.ini
@@ -65,7 +65,7 @@ deps =
     -r{toxinidir}/docs/requirements.txt
 commands =
     sphinx-build {posargs:-E} -b html docs dist/docs
-    sphinx-build -b linkcheck docs dist/docs
+    #sphinx-build -b linkcheck docs dist/docs
 
 [testenv:coveralls]
 deps =
